added dagger core module for DI graph added home feature to show home added a dummy dynamic feature to see how they work
