+++ app/src/main/java/com/worldsnas/aio/AIOApp.kt
@@ -1,16 +1,26 @@
 package com.worldsnas.aio
 
 import android.app.Application
-import com.worldsnas.aio.di.DaggerAppComponent
-import com.worldsnas.core.di.CoreComponent
-import com.worldsnas.core.di.CoreComponentProvider
+import com.worldsnas.aio.BuildConfig.DEBUG
+import com.worldsnas.daggercore.CoreComponent
+import com.worldsnas.daggercore.CoreComponentProvider
+import com.worldsnas.daggercore.DaggerCoreComponent
+import timber.log.Timber
 
 class AIOApp : Application(), CoreComponentProvider {
 
     private val coreComponent by lazy {
-        DaggerAppComponent.builder().build()
+        DaggerCoreComponent.builder().build()
     }
 
     override fun core(): CoreComponent =
         coreComponent
+
+    override fun onCreate() {
+        super.onCreate()
+
+        if (DEBUG) {
+            Timber.plant(Timber.DebugTree())
+        }
+    }
 }
