--- app/src/main/java/com/worldsnas/aio/home/HomeIntent.kt
@@ -1,7 +1,0 @@
-package com.worldsnas.aio.home
-
-import com.worldsnas.mvi.MviIntent
-
-sealed class HomeIntent : MviIntent {
-    object Initial : HomeIntent()
-}
