+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">AIO</string>
+    <string name="title_domy">domy</string>
 </resources>
