+++ core/src/main/java/com/worldsnas/core/ViewExtensions.kt
@@ -1,8 +1,6 @@
 package com.worldsnas.core
 
 import android.view.View
-import com.worldsnas.core.di.CoreComponent
-import com.worldsnas.core.di.CoreComponentProvider
 
 @Suppress("LiftReturnOrAssignment")
 infix fun View.visible(visible: Boolean) =
@@ -10,11 +8,4 @@ infix fun View.visible(visible: Boolean) =
         visibility = View.VISIBLE
     } else {
         visibility = View.GONE
-    }
-
-fun View.coreComponent(): CoreComponent =
-    if (context.applicationContext is CoreComponentProvider) {
-        (context.applicationContext as CoreComponentProvider).core()
-    } else {
-        throw IllegalArgumentException("app class must implement CoreComponentProvider")
     }
