--- core/src/main/java/com/worldsnas/core/di/CoreComponentProvider.kt
@@ -1,6 +1,0 @@
-package com.worldsnas.core.di
-
-interface CoreComponentProvider {
-
-    fun core(): CoreComponent
-}
