--- core/src/main/java/com/worldsnas/core/di/modules/DatabaseModule.kt
@@ -1,7 +1,0 @@
-package com.worldsnas.core.di.modules
-
-import dagger.Module
-
-@Module
-object DatabaseModule {
-}
