--- core/src/main/java/com/worldsnas/core/di/modules/network/EndPoint.kt
@@ -1,4 +1,0 @@
-package com.worldsnas.core.di.modules.network
-
-
-const val BASE_URL = "https://stage.mobank.co"
