--- core/src/main/java/com/worldsnas/core/di/scope/AppScope.kt
@@ -1,7 +1,0 @@
-package com.worldsnas.core.di.scope
-
-import javax.inject.Scope
-
-@Scope
-@Retention(AnnotationRetention.RUNTIME)
-annotation class AppScope
