+++ daggercore/src/main/java/com/worldsnas/daggercore/CoreComponentProvider.kt
@@ -1,0 +1,6 @@
+package com.worldsnas.daggercore
+
+interface CoreComponentProvider {
+
+    fun core(): CoreComponent
+}
