+++ daggercore/src/main/java/com/worldsnas/daggercore/modules/DatabaseModule.kt
@@ -1,0 +1,7 @@
+package com.worldsnas.daggercore.modules
+
+import dagger.Module
+
+@Module
+object DatabaseModule {
+}
