+++ daggercore/src/main/java/com/worldsnas/daggercore/modules/network/EndPoint.kt
@@ -1,0 +1,4 @@
+package com.worldsnas.daggercore.modules.network
+
+
+const val BASE_URL = "https://stage.mobank.co"
