+++ daggercore/src/main/java/com/worldsnas/daggercore/scope/ActivityScope.kt
@@ -1,0 +1,7 @@
+package com.worldsnas.daggercore.scope
+
+import javax.inject.Scope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class ActivityScope
