+++ home/src/main/java/com/worldsnas/home/HomeIntent.kt
@@ -1,0 +1,7 @@
+package com.worldsnas.home
+
+import com.worldsnas.mvi.MviIntent
+
+sealed class HomeIntent : MviIntent {
+    object Initial : HomeIntent()
+}
