+++ home/src/main/java/com/worldsnas/home/HomeState.kt
@@ -1,0 +1,12 @@
+package com.worldsnas.home
+
+import com.worldsnas.base.BaseState
+import com.worldsnas.base.BaseViewState
+
+data class HomeState(
+    override val base: BaseState
+) : BaseViewState {
+    companion object {
+        fun startingState() = HomeState(base = BaseState.showError("اصغر"))
+    }
+}
