+++ home/src/main/java/com/worldsnas/home/di/HomeComponent.kt
@@ -1,0 +1,12 @@
+package com.worldsnas.home.di
+
+import com.worldsnas.daggercore.CoreComponent
+import com.worldsnas.daggercore.scope.FeatureScope
+import com.worldsnas.home.HomeView
+import dagger.Component
+
+@FeatureScope
+@Component(dependencies = [CoreComponent::class], modules = [HomeModule::class])
+interface HomeComponent{
+    fun inject(view : HomeView)
+}
