+++ home/src/main/res/layout/view_home.xml
