+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':base', ':mvi', ':core', ':domain', ':panther'
+include ':app', ':base', ':mvi', ':core', ':domain', ':panther', ':domy', ':daggercore', ':home'
